--- src/main/java/com/truve/platform/auth/service/common/constants/UserRole.java
@@ -1,8 +1,0 @@
-package com.truve.platform.auth.service.common.constants;
-
-public enum UserRole {
-	ADMIN,
-	MEMBER,
-	ARTIST,
-	;
-}
