+++ src/main/java/com/truve/platform/user/service/UserServiceApplication.java
@@ -1,13 +1,13 @@
-package com.truve.platform.auth.service;
+package com.truve.platform.user.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AuthServiceApplication {
+public class UserServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AuthServiceApplication.class, args);
+		SpringApplication.run(UserServiceApplication.class, args);
 	}
 
 }
