+++ src/main/java/com/truve/platform/user/service/common/constants/UserRole.java
@@ -1,0 +1,8 @@
+package com.truve.platform.user.service.common.constants;
+
+public enum UserRole {
+	ADMIN,
+	MEMBER,
+	ARTIST,
+	;
+}
