+++ src/main/java/com/truve/platform/user/service/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.truve.platform.auth.service.common.exception;
+package com.truve.platform.user.service.common.exception;
 
 import org.springframework.http.HttpStatus;
