+++ src/main/java/com/truve/platform/user/service/domain/repository/User.java
@@ -1,21 +1,21 @@
-package com.truve.platform.auth.service.domain.repository;
+package com.truve.platform.user.service.domain.repository;
 
-import com.truve.platform.auth.service.common.constants.AuthProvider;
-import com.truve.platform.auth.service.common.constants.UserRole;
-import com.truve.platform.auth.service.common.support.BaseEntity;
+import com.truve.platform.user.service.common.constants.AuthProvider;
+import com.truve.platform.user.service.common.constants.UserRole;
+import com.truve.platform.user.service.common.support.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "users")
 public class User extends BaseEntity {
