+++ src/test/java/com/truve/platform/user/service/UserServiceApplicationTests.java
@@ -1,10 +1,10 @@
-package com.truve.platform.auth.service;
+package com.truve.platform.user.service;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AuthServiceApplicationTests {
+class UserServiceApplicationTests {
 
 	@Test
 	void contextLoads() {
